script: Move all gamepad functionality into Window.
